+++ images/minimal-notebook/setup-scripts/setup_julia.py
@@ -43,12 +43,6 @@ def get_latest_julia_url() -> tuple[str, str]:
     triplet = unify_aarch64(platform.machine()) + "-linux-gnu"
     file_info = [vf for vf in latest_version_files if vf["triplet"] == triplet][0]
     LOGGER.info(f"Latest version: {file_info['version']} url: {file_info['url']}")
-    BROKEN_VERSION = "1.11.3"
-    if file_info["version"] == BROKEN_VERSION:
-        LOGGER.warning(
-            f"Not using Julia {BROKEN_VERSION}, because it hangs in GitHub self-hosted runners"
-        )
-        return file_info["url"].replace(BROKEN_VERSION, "1.11.1"), "1.11.1"
     return file_info["url"], file_info["version"]